`IndexedDbQuery` and fixed circular imports
